moved user-from-token detection to appInstance
