+++ src/application-instance.ts
@@ -1,6 +1,10 @@
+import { decode } from "jsonwebtoken";
+
 import Application from "./application";
 import Resource from "./resource";
-import { OperationProcessor } from ".";
+import OperationProcessor from "./processors/operation-processor";
+import { Operation } from "./types";
+import jsonApiErrors from "./json-api-errors";
 
 export default class ApplicationInstance {
   public user: typeof Resource;
@@ -11,5 +15,35 @@ export default class ApplicationInstance {
     resourceType: string
   ): Promise<OperationProcessor<Resource> | undefined> {
     return this.app.processorFor(resourceType, this);
+  }
+
+  async getUserFromToken(token: string): Promise<typeof Resource | undefined> {
+    const tokenPayload = decode(token);
+
+    if (!tokenPayload) {
+      throw jsonApiErrors.InvalidToken();
+    }
+
+    const userId = tokenPayload["id"];
+
+    if (!userId) return;
+
+    const op = {
+      op: "identify",
+      ref: {
+        type: "user",
+        id: userId
+      },
+      params: {}
+    } as Operation;
+
+    const processor = await this.processorFor(op.ref.type);
+
+    if (!processor) {
+      return;
+    }
+
+    const user = await this.app.executeOperation(op, processor);
+    return user.data[0];
   }
 }
